Refactor module layout
Rename container mod to npk. Make cgroups available on Linux only.
