--- north/src/cgroups/mod.rs
@@ -1,23 +1,0 @@
-// Copyright (c) 2019 - 2020 ESRLabs
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-mod target;
-#[cfg(any(target_os = "android", target_os = "linux"))]
-pub use target::*;
-
-#[cfg(not(any(target_os = "android", target_os = "linux")))]
-mod mock;
-#[cfg(not(any(target_os = "android", target_os = "linux")))]
-pub use mock::*;
