--- north/src/container/mod.rs
@@ -1,36 +1,0 @@
-// Copyright (c) 2019 - 2020 ESRLabs
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-
-use async_std::path::PathBuf;
-use north_common::manifest::Manifest;
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-mod linux;
-#[cfg(any(target_os = "android", target_os = "linux"))]
-pub use linux::{install, install_all, uninstall};
-
-#[cfg(not(any(target_os = "android", target_os = "linux")))]
-mod osx;
-#[cfg(not(any(target_os = "android", target_os = "linux")))]
-pub use osx::{install, install_all, uninstall};
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-mod hashes;
-
-#[derive(Debug)]
-pub struct Container {
-    pub manifest: Manifest,
-    pub root: PathBuf,
-    pub data: PathBuf,
-}
